[dnm] drivers: pinctrl: another take on clockpin
A solution that skips users dealing with the CLOCKPIN setting. This is
still subobtimal: nordic,clock-enable is stored on every single pin
entry.

We may go to the previous approach by hardcoding stuff in the driver.

Let me know about your thoughts/ideas, this is the _best_ I could came
up this morning.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>