drivers: nvmem_provider: introduce new NVMEM provider class
While we have a NVMEM consumer subsystem, there's nothing for providers.
The consumer subsystem just supports EEPROM devices as a provider,
however, other common cases like EFUSEs are not covered anywhere. It
looks like there's a syscon device for Bouffalo efuses, but it feels
like an abuse of the syscon API as a sort of "catch-all" type of API.

This patch is inspired by the Linux NVMEM consumer APIs, always of
course in a much simple manner (just read/write for now).

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>